requirements.txt
avro==1.10.0
azure-core==1.23.0
azure-eventhub==5.7.0
azure-identity==1.8.0
azure-schemaregistry==1.0.0
azure-schemaregistry-avroserializer==1.0.0b4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.2
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
idna==3.3
isodate==0.6.1
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
msal==1.17.0
msal-extensions==0.3.1
msrest==0.6.21
numpy==1.22.3
oauthlib==3.2.0
portalocker==2.4.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==1.4.32
typing_extensions==4.1.1
uamqp==1.5.2
urllib3==1.26.9
Werkzeug==2.0.3
